requirements.txt
# ------------------------
# Web / API
# ------------------------
fastapi
uvicorn[standard]

# ------------------------
# Database
# ------------------------
sqlalchemy>=2.0
asyncpg
alembic

# ------------------------
# Auth & Security
# ------------------------
python-jose
passlib[bcrypt]
bcrypt<4.0.0
pydantic
pydantic-settings
email-validator

# ------------------------
# HTTP client (parsing)
# ------------------------
httpx

# ------------------------
# ML / Recommender
# ------------------------
scikit-learn
sentence-transformers
numpy
scipy

# ------------------------
# Utils
# ------------------------
python-dotenv
psycopg2-binary
python-multipart
pillow
